+++ src/commands/submit.ts
@@ -6,7 +6,7 @@ import { leetCodeTreeDataProvider } from "../explorer/LeetCodeTreeDataProvider";
 import { leetCodeExecutor } from "../leetCodeExecutor";
 import { leetCodeManager } from "../leetCodeManager";
 import { DialogType, promptForOpenOutputChannel, promptForSignIn } from "../utils/uiUtils";
-import { getActiveFilePath } from "../utils/workspaceUtils";
+import { getActiveSolutionFile, IActiveSolutionFile } from "../utils/workspaceUtils";
 import { leetCodeSubmissionProvider } from "../webview/leetCodeSubmissionProvider";
 
 export async function submitSolution(uri?: vscode.Uri): Promise<void> {
@@ -15,17 +15,19 @@ export async function submitSolution(uri?: vscode.Uri): Promise<void> {
         return;
     }
 
-    const filePath: string | undefined = await getActiveFilePath(uri);
-    if (!filePath) {
+    const solutionFile: IActiveSolutionFile | undefined = await getActiveSolutionFile(uri);
+    if (!solutionFile) {
         return;
     }
 
     try {
-        const result: string = await leetCodeExecutor.submitSolution(filePath);
+        const result: string = await leetCodeExecutor.submitSolution(solutionFile.filePath);
         leetCodeSubmissionProvider.show(result);
     } catch (error) {
         await promptForOpenOutputChannel("Failed to submit the solution. Please open the output channel for details.", DialogType.error);
         return;
+    } finally {
+        await solutionFile.dispose();
     }
 
     leetCodeTreeDataProvider.refresh();