+++ src/commands/test.ts
@@ -8,18 +8,19 @@ import { leetCodeManager } from "../leetCodeManager";
 import { IQuickItemEx, UserStatus } from "../shared";
 import { isWindows, usingCmd } from "../utils/osUtils";
 import { DialogType, promptForOpenOutputChannel, showFileSelectDialog } from "../utils/uiUtils";
-import { getActiveFilePath } from "../utils/workspaceUtils";
+import { getActiveSolutionFile, IActiveSolutionFile } from "../utils/workspaceUtils";
 import * as wsl from "../utils/wslUtils";
 import { leetCodeSubmissionProvider } from "../webview/leetCodeSubmissionProvider";
 
 export async function testSolution(uri?: vscode.Uri): Promise<void> {
+    let solutionFile: IActiveSolutionFile | undefined;
     try {
         if (leetCodeManager.getStatus() === UserStatus.SignedOut) {
             return;
         }
 
-        const filePath: string | undefined = await getActiveFilePath(uri);
-        if (!filePath) {
+        solutionFile = await getActiveSolutionFile(uri);
+        if (!solutionFile) {
             return;
         }
         const picks: Array<IQuickItemEx<string>> = [];
@@ -51,7 +52,7 @@ export async function testSolution(uri?: vscode.Uri): Promise<void> {
         let result: string | undefined;
         switch (choice.value) {
             case ":default":
-                result = await leetCodeExecutor.testSolution(filePath);
+                result = await leetCodeExecutor.testSolution(solutionFile.filePath);
                 break;
             case ":direct":
                 const testString: string | undefined = await vscode.window.showInputBox({
@@ -61,15 +62,15 @@ export async function testSolution(uri?: vscode.Uri): Promise<void> {
                     ignoreFocusOut: true,
                 });
                 if (testString) {
-                    result = await leetCodeExecutor.testSolution(filePath, parseTestString(testString));
+                    result = await leetCodeExecutor.testSolution(solutionFile.filePath, parseTestString(testString));
                 }
                 break;
             case ":file":
-                const testFile: vscode.Uri[] | undefined = await showFileSelectDialog(filePath);
+                const testFile: vscode.Uri[] | undefined = await showFileSelectDialog(solutionFile.filePath);
                 if (testFile && testFile.length) {
                     const input: string = (await fse.readFile(testFile[0].fsPath, "utf-8")).trim();
                     if (input) {
-                        result = await leetCodeExecutor.testSolution(filePath, parseTestString(input.replace(/\r?\n/g, "\\n")));
+                        result = await leetCodeExecutor.testSolution(solutionFile.filePath, parseTestString(input.replace(/\r?\n/g, "\\n")));
                     } else {
                         vscode.window.showErrorMessage("The selected test file must not be empty.");
                     }
@@ -84,6 +85,10 @@ export async function testSolution(uri?: vscode.Uri): Promise<void> {
         leetCodeSubmissionProvider.show(result);
     } catch (error) {
         await promptForOpenOutputChannel("Failed to test the solution. Please open the output channel for details.", DialogType.error);
+    } finally {
+        if (solutionFile) {
+            await solutionFile.dispose();
+        }
     }
 }