+++ src/leetCodeExecutor.ts
@@ -101,6 +101,24 @@ class LeetCodeExecutor implements Disposable {
     }
 
     public async showProblem(problemNode: IProblem, language: string, filePath: string, showDescriptionInComment: boolean = false, needTranslation: boolean): Promise<void> {
+        if (!await fse.pathExists(filePath)) {
+            await fse.createFile(filePath);
+            const codeTemplate: string = await this.getProblemTemplate(
+                problemNode,
+                language,
+                showDescriptionInComment,
+                needTranslation,
+            );
+            await fse.writeFile(filePath, codeTemplate);
+        }
+    }
+
+    public async getProblemTemplate(
+        problemNode: IProblem,
+        language: string,
+        showDescriptionInComment: boolean = false,
+        needTranslation: boolean,
+    ): Promise<string> {
         const templateType: string = showDescriptionInComment ? "-cx" : "-c";
         const cmd: string[] = [await this.getLeetCodeBinaryPath(), "show", problemNode.id, templateType, "-l", language];
 
@@ -108,11 +126,7 @@ class LeetCodeExecutor implements Disposable {
             cmd.push("-T"); // use -T to force English version
         }
 
-        if (!await fse.pathExists(filePath)) {
-            await fse.createFile(filePath);
-            const codeTemplate: string = await this.executeCommandWithProgressEx("Fetching problem data...", this.nodeExecutable, cmd);
-            await fse.writeFile(filePath, codeTemplate);
-        }
+        return this.executeCommandWithProgressEx("Fetching problem data...", this.nodeExecutable, cmd);
     }
 
     /**