+++ src/utils/httpUtils.ts
@@ -5,8 +5,8 @@ import { DialogType, promptForOpenOutputChannel } from "./uiUtils";
 
 const referer = "vscode-lc-extension";
 
-export function LcAxios<T = any>(path: string, settings?: AxiosRequestConfig): AxiosPromise<T> {
-    const cookie = globalState.getCookie();
+export function LcAxios<T = any>(path: string, settings?: AxiosRequestConfig, cookieOverride?: string): AxiosPromise<T> {
+    const cookie = cookieOverride || globalState.getCookie();
     if (!cookie) {
         promptForOpenOutputChannel(
             `Failed to obtain the cookie. Please log in again.`,